Move render_abstract_page to views.py
