+++ indico/modules/events/abstracts/views.py
@@ -16,6 +16,8 @@
 
 from __future__ import unicode_literals
 
+from flask import render_template, session
+
 from MaKaC.common.TemplateExec import render
 from MaKaC.webinterface.pages.base import WPJinjaMixin
 from MaKaC.webinterface.pages.conferences import WPConferenceDefaultDisplayBase, WPConferenceModifBase
@@ -92,3 +94,24 @@ class WPDisplayAbstractsReviewing(WPDisplayAbstracts):
         return (WPDisplayAbstracts.getCSSFiles(self) +
                 self._asset_env['event_display_sass'].urls() +
                 self._asset_env['tracks_sass'].urls())
+
+
+def render_abstract_page(abstract, view_class=None, management=False):
+    from indico.modules.events.abstracts.forms import (AbstractCommentForm, AbstractJudgmentForm,
+                                                       AbstractReviewedForTracksForm, build_review_form)
+    comment_form = AbstractCommentForm(abstract=abstract, user=session.user, formdata=None)
+    review_forms = {track.id: build_review_form(abstract, track)
+                    for track in abstract.reviewed_for_tracks
+                    if track.can_review_abstracts(session.user)}
+    judgment_form = AbstractJudgmentForm(abstract=abstract, formdata=None)
+    review_track_list_form = AbstractReviewedForTracksForm(event=abstract.event_new, obj=abstract, formdata=None)
+    params = {'abstract': abstract,
+              'comment_form': comment_form,
+              'review_forms': review_forms,
+              'review_track_list_form': review_track_list_form,
+              'judgment_form': judgment_form,
+              'management': management}
+    if view_class:
+        return view_class.render_template('abstract.html', abstract.event_new.as_legacy, **params)
+    else:
+        return render_template('events/abstracts/abstract.html', no_javascript=True, **params)
